Move mock client to php-http/mock-client
